dev_env/linebot-pipeline/requirements.txt
langchain==0.0.352
openai
sentence_transformers
faiss-cpu
line-bot-sdk
python-dotenv